pyproject.toml
[tool.poetry]
name = "uselessapi"
version = "0.1.0"
description = ""
authors = ["winter_x64 <itswinterx64@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
supabase = "^2.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
